+++ benchopt/plotting/generate_html.py
@@ -107,6 +107,7 @@ def get_results(fnames, kinds, root_html, benchmark_name, copy=False,
             obj_cols=[k for k in df.columns if k.startswith('objective_')
                       and k != 'objective_name'],
             kinds=list(kinds),
+            metadata=get_metadata(df),
         )
 
         # JSON
@@ -124,6 +125,43 @@ def get_results(fnames, kinds, root_html, benchmark_name, copy=False,
         )
 
     return results
+
+
+def get_metadata(df):
+    """Get the benchmark metadata.
+
+    Metadata are already available among the columns of `df`.
+    It might be Objective and/or Solvers description.
+
+    Returns
+    -------
+    metadata: dict
+        Dictionary containing the benchmark metadata.
+    """
+    metadata = {}
+
+    # get solver descriptions
+    # wrap in try-except block to preserve compatibility
+    # with older versions
+    try:
+        solvers_description = df.groupby(
+            by=["solver_name"]
+        )["solver_description"].first()
+
+        metadata["solvers_description"] = solvers_description.to_dict()
+    except KeyError:
+        metadata["solvers_description"] = {}
+
+    # to avoid conflicts with objective metrics
+    # get objective description and use `obj_` instead of `objective_`
+    # try-except block to preserve compatibility with benchopt <= v1.3.1
+    try:
+        obj_description = df["obj_description"].unique()[0]
+        metadata["obj_description"] = obj_description
+    except KeyError:
+        metadata["obj_description"] = ""
+
+    return metadata
 
 
 def shape_datasets_for_html(df, plot_config=None):
@@ -183,6 +221,8 @@ def shape_solvers_for_html(df, objective_column):
         # remove infinite values
         df_filtered = df_filtered.replace([np.inf, -np.inf], np.nan)
         df_filtered = df_filtered.dropna(subset=[objective_column])
+        if len(df_filtered) == 0:
+            continue
 
         # compute median of 'time' and objective_column
         fields = ["time", objective_column]