+++ benchopt/tests/test_benchmarks/dummy_benchmark/solvers/solver_test.py
@@ -52,6 +52,5 @@ class Solver(BaseSolver):
     def get_result(self):
         return self.w
 
-    @staticmethod
-    def get_next(stop_val):
+    def get_next(self, stop_val):
         return stop_val + 1